Move application infra to new package
