+++ infra/src/main/kotlin/uk/co/developmentanddinosaurs/apps/poker/infra/application/PokerEc2.kt
@@ -1,4 +1,4 @@
-package uk.co.developmentanddinosaurs.apps.poker.infra
+package uk.co.developmentanddinosaurs.apps.poker.infra.application
 
 import software.amazon.awscdk.services.ec2.Instance
 import software.amazon.awscdk.services.ec2.InstanceClass
@@ -6,16 +6,14 @@ import software.amazon.awscdk.services.ec2.InstanceSize
 import software.amazon.awscdk.services.ec2.InstanceType
 import software.amazon.awscdk.services.ec2.MachineImage
 import software.amazon.awscdk.services.ec2.MultipartUserData
-import software.amazon.awscdk.services.ec2.Peer
-import software.amazon.awscdk.services.ec2.Port
 import software.amazon.awscdk.services.ec2.S3DownloadOptions
-import software.amazon.awscdk.services.ec2.SecurityGroup
 import software.amazon.awscdk.services.ec2.SubnetSelection
 import software.amazon.awscdk.services.ec2.SubnetType
 import software.amazon.awscdk.services.ec2.UserData
 import software.amazon.awscdk.services.iam.Role
 import software.amazon.awscdk.services.s3.Bucket
 import software.constructs.Construct
+import uk.co.developmentanddinosaurs.apps.poker.infra.common.PokerBucket
 
 class PokerEc2(private val scope: Construct) {
     fun instance(): Instance {
@@ -24,15 +22,8 @@ class PokerEc2(private val scope: Construct) {
             .role(Role.fromRoleName(scope, "PlanningPokerApplication", "PlanningPokerApplication"))
             .machineImage(MachineImage.latestAmazonLinux2023()).vpc(PokerVpc.default(scope))
             .vpcSubnets(SubnetSelection.builder().subnetType(SubnetType.PUBLIC).build()).keyName("planning-poker")
-            .securityGroup(securityGroup()).userData(userData()).associatePublicIpAddress(true).build()
-    }
-
-    private fun securityGroup(): SecurityGroup {
-        val securityGroup =
-            SecurityGroup.Builder.create(scope, "PokerSecurityGroup").securityGroupName("planning-poker-sg")
-                .description("Planning Poker Security Group").vpc(PokerVpc.default(scope)).build()
-        securityGroup.addIngressRule(Peer.anyIpv4(), Port.allTraffic())
-        return securityGroup
+            .securityGroup(PokerSecurityGroup(scope).securityGroup()).userData(userData())
+            .associatePublicIpAddress(true).build()
     }
 
     private fun userData(): UserData {
@@ -42,7 +33,7 @@ class PokerEc2(private val scope: Construct) {
         initConfig.addCommands("- [scripts-user, always]")
         val userData = UserData.forLinux()
         userData.addCommands("sudo yum install -y java-17-amazon-corretto-headless")
-        userData.addS3DownloadCommand(download("planning-poker-application", "planning-poker.jar"))
+        userData.addS3DownloadCommand(download(PokerBucket(scope).bucketReference().bucketName, "planning-poker.jar"))
         userData.addCommands("sudo java -jar /app/planning-poker.jar &")
         val multipartUserData = MultipartUserData()
         multipartUserData.addUserDataPart(initConfig)
@@ -50,7 +41,10 @@ class PokerEc2(private val scope: Construct) {
         return multipartUserData
     }
 
-    private fun download(bucket: String, key: String): S3DownloadOptions {
+    private fun download(
+        bucket: String,
+        key: String,
+    ): S3DownloadOptions {
         return S3DownloadOptions.builder().bucket(Bucket.fromBucketName(scope, "bucket", bucket)).localFile("/app/$key")
             .bucketKey(key).build()
     }
