+++ infra/src/main/kotlin/uk/co/developmentanddinosaurs/apps/poker/infra/application/PokerVpc.kt
@@ -1,4 +1,4 @@
-package uk.co.developmentanddinosaurs.apps.poker.infra
+package uk.co.developmentanddinosaurs.apps.poker.infra.application
 
 import software.amazon.awscdk.services.ec2.IVpc
 import software.amazon.awscdk.services.ec2.VpcLookupOptions
@@ -8,6 +8,7 @@ import software.amazon.awscdk.services.ec2.Vpc as AwsVpc
 class PokerVpc {
     companion object {
         private var vpc: IVpc? = null
+
         fun default(construct: Construct): IVpc {
             if (vpc != null) return vpc!!
             vpc = AwsVpc.fromLookup(construct, "DefaultVpc", VpcLookupOptions.builder().isDefault(true).build())
